Rewrite import code to fix multiple issues
This new version is simpler than the old wrappers. It supports almost all of
the properties that can be saved. Support for importing alarms is almost
complete. All day and recurring events are properly supported as are
availability, organiser etc. This should fix github issues 1, 3, 14 and 15.

Event deletion is disabled temporarily since it needs refactoring that would
conflict with these changes. Also the behaviour of Timezones under the new
code needs investigation.