Fix duplicate handling for Issue #23
Add an option for UID's to be global across all calendars, defaulting to off.
Add another duplicate handling option to replace duplicate events in the
calendar they came from (this is not the default).

Although UIDs are meant to be globally unique, it can be counter-intuitive
to have events in other calendars changed without having configured the app
to do so.

This change reworks the processing code to hide some of the added complexity,
and prevents the query for duplicates being run twice. It also fixes the
id types from int to long.