Enable debug logging option for users
We will need extended logging to debug user issues.

When enabled, verbose and debug messages are output as information messages
so they are not filtered nor is the code stripped out at runtime. In debug
builds the option has no effect.

Also information level logging was being filtered in debug builds which is
counter to the logging documentation on usage. Information messages are
now always output.