Initial, fairly minimal support for import testing
This works by using events containing an X-TEST-NAME and then
one or more X-TEST-VALUE properties. The X-TEST-VALUE properties are
checked against the resulting ContentValues that (in non-test mode)
would be inserted into the DB.

This approach has its limitations but it allows us to start testing
the many permutations of inputs and to prevent regressions. The test
support can be incrementally improved as tests are added.