Fix organizer handling, add test.
This fixes the parsing and output of organizer so that it contains just the
actual to: part of the mailto: URI. This should fix the value in the Android
DB and so fix any check where the calendar owner is compared to the
organizer. It doesn't tell us why GUESTS_CAN_MODIFY isn't working however.
Lets see if this is enough to fix #29 for the reporting user.