Try again with a null projection if the provider throws an NPE
It seems that either one of the query column names has become null (the docs
on the API level supporting it, and thus our backwards compatiblity shim are
wrong), or some phone makers have customised the calendar store and removed
columns that the query is meant to provide (perhaps failing to keep up with
underlying AOSP improvements to the calendar model).

Add some logging to differentiate these cases. If the former case holds, we
should be able to fix the shim (good since its used by others).

Hopefully fixes #33.